+++ CHANGELOG.md
@@ -26,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The scroll compensation isn't applied now when the window scrollY position is 0.
   So when some content appears on the top of the page, it will not caused scroll
   compensation and will not be hidden under the top window edge.
+- The info popups now shows properly for users that changed their usernames.
 
 ### Changed
 - Removed old invalid links to Clio and "Archives F.A.Q."